REX-848: Moved test
